+++ tests/PureLogicTests.cs
@@ -2794,12 +2794,50 @@ namespace AshAndEmber.Tests
         }
 
         [Test]
-        public void DemonMath_CastsMagic_OnlyRavager()
+        public void DemonMath_CastsMagic_RavagerAndLordOnly()
         {
             Assert.IsFalse(DemonMath.CastsMagic(DemonMath.DemonTier.Fiend));
             Assert.IsFalse(DemonMath.CastsMagic(DemonMath.DemonTier.Stalker));
             Assert.IsTrue(DemonMath.CastsMagic(DemonMath.DemonTier.Ravager));
             Assert.IsFalse(DemonMath.CastsMagic(DemonMath.DemonTier.Hellsteed));
+            Assert.IsTrue(DemonMath.CastsMagic(DemonMath.DemonTier.Lord));
+        }
+
+        [Test]
+        public void DemonMath_LordCastPatternIndex_AlternatesFireThenSpirit()
+        {
+            Assert.AreEqual(0, DemonMath.LordCastPatternIndex(0));
+            Assert.AreEqual(1, DemonMath.LordCastPatternIndex(1));
+            Assert.AreEqual(0, DemonMath.LordCastPatternIndex(2));
+            Assert.AreEqual(1, DemonMath.LordCastPatternIndex(3));
+            Assert.AreEqual(0, DemonMath.LordCastPatternIndex(-1)); // never throws on a bad count
+        }
+
+        [Test]
+        public void DemonMath_VisualScale_RavagerAndLordAreBiggestFiendUnchanged()
+        {
+            Assert.AreEqual(1.00f, DemonMath.VisualScale(DemonMath.DemonTier.Fiend), 0.001f);
+            Assert.Greater(DemonMath.VisualScale(DemonMath.DemonTier.Stalker), 1.00f);
+            Assert.Greater(DemonMath.VisualScale(DemonMath.DemonTier.Ravager), DemonMath.VisualScale(DemonMath.DemonTier.Stalker));
+            Assert.Greater(DemonMath.VisualScale(DemonMath.DemonTier.Lord), DemonMath.VisualScale(DemonMath.DemonTier.Ravager));
+        }
+
+        [Test]
+        public void DemonMath_MonsterIdFor_OnlyRavagerAndLordGetTheHulkingCapsule()
+        {
+            Assert.IsNull(DemonMath.MonsterIdFor(DemonMath.DemonTier.Fiend));
+            Assert.IsNull(DemonMath.MonsterIdFor(DemonMath.DemonTier.Stalker));
+            Assert.AreEqual(DemonMath.HulkingMonsterId, DemonMath.MonsterIdFor(DemonMath.DemonTier.Ravager));
+            Assert.IsNull(DemonMath.MonsterIdFor(DemonMath.DemonTier.Hellsteed));
+            Assert.AreEqual(DemonMath.HulkingMonsterId, DemonMath.MonsterIdFor(DemonMath.DemonTier.Lord));
+        }
+
+        [Test]
+        public void DemonMath_SpeedMultiplier_FiendAndStalkerFasterRavagerSlower()
+        {
+            Assert.Greater(DemonMath.SpeedMultiplier(DemonMath.DemonTier.Fiend), 1.00f);
+            Assert.Greater(DemonMath.SpeedMultiplier(DemonMath.DemonTier.Stalker), DemonMath.SpeedMultiplier(DemonMath.DemonTier.Fiend));
+            Assert.Less(DemonMath.SpeedMultiplier(DemonMath.DemonTier.Ravager), 1.00f);
         }
 
         [Test]